+++ src/campominado/Hibernate.java
@@ -1,3 +1,4 @@
+package campominado;
 import org.hibernate.SessionFactory;
 import org.hibernate.boot.MetadataSources;
 import org.hibernate.boot.registry.StandardServiceRegistry;
