+++ src/campominado/Historico.java
@@ -1,0 +1,5 @@
+package campominado;
+
+public class Historico {
+
+}
