+++ src/campominado/Main.java
@@ -1,7 +1,8 @@
+package campominado;
 import javax.swing.SwingUtilities;
 
 public class Main {
-	
+
 	public static void main(String[] args) {
 		
 		SwingUtilities.invokeLater(new Runnable() {
@@ -12,6 +13,7 @@ public class Main {
 			}
 			
 		});
+
 	}
 
 }
