+++ src/campominado/TelaHistorico.java
@@ -1,3 +1,4 @@
+package campominado;
 import javax.swing.JFrame;
 import java.awt.FlowLayout;
