+++ src/campominado/TelaInicial.java
@@ -1,3 +1,4 @@
+package campominado;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.JFrame;
